requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.1
django-crontab==0.7.1
django-dotenv==1.4.2
djangorestframework==3.13.1
drf-nested-routers==0.93.4
kombu==5.2.4
packaging==21.3
pip-autoremove==0.10.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycodestyle==2.9.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
